# api/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Heimdallr API
  version: 1.0.0
  description: API for managing applications, releases, compliance reports, servers, agents, providers, automations, and automation jobs.
servers:
  - url: /api
    description: API base path
...
security:
  - BearerAuth: []
  - SessionCookie: []
paths:
...
  /v1/application/{application_id}/release/{release_id}/report:
    get:
      tags:
        - Report
      summary: List reports for a release
      operationId: listReleaseReports
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - $ref: "#/components/parameters/ReleaseIDPath"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Reports returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Report
      summary: Create report for a release
      operationId: createReleaseReport
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - $ref: "#/components/parameters/ReleaseIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateRequest"
      responses:
        "201":
          description: Report created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/application/{application_id}/release/{release_id}/report/{report_id}:
    get:
      tags:
        - Report
      summary: Get report by ID
      operationId: getReleaseReport
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - $ref: "#/components/parameters/ReleaseIDPath"
        - $ref: "#/components/parameters/ReportIDPath"
      responses:
        "200":
          description: Report returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

    patch:
      tags:
        - Report
      summary: Update report status
      operationId: updateReleaseReport
      parameters:
        - $ref: "#/components/parameters/ApplicationIDPath"
        - $ref: "#/components/parameters/ReleaseIDPath"
        - $ref: "#/components/parameters/ReportIDPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdateRequest"
      responses:
        "200":
          description: Report updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportDataResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /v1/report:
    get:
      tags:
        - Report
      summary: List reports across applications
      operationId: listReportsGlobal
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: application_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: release_id
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/UUID"
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - started
              - skipped
              - success
              - failed
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum:
              - sast
              - dast
              - sbom
              - code_coverage
              - custom
      responses:
        "200":
          description: Reports returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API token passed as a Bearer token in the Authorization header
    SessionCookie:
      type: apiKey
      in: cookie
      name: heimdallr_session
      description: HttpOnly browser session cookie
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    ApplicationIDPath:
      name: application_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReleaseIDPath:
      name: release_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ReportIDPath:
      name: report_id
      in: path
      required: true
      schema:
        type: string
        minLength: 1
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    UUID:
      type: string
      format: uuid
      x-go-type: uuid.UUID
      x-go-type-import:
        path: github.com/google/uuid
        name: uuid
    URL:
      type: string
      format: uri
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: invalid request body
    Pagination:
      type: object
      required:
        - page
        - limit
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        total_pages:
          type: integer
          minimum: 0
...
    JobStatus:
      type: string
      enum:
        - started
        - skipped
        - success
        - failed
...
    Report:
      type: object
      required:
        - id
        - application_id
        - release_id
        - application
        - version
        - type
        - status
        - created_at
      properties:
        id:
          type: string
        application_id:
          $ref: "#/components/schemas/UUID"
        release_id:
          $ref: "#/components/schemas/UUID"
        application:
          type: string
        version:
          type: string
        type:
          $ref: "#/components/schemas/ReportType"
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
        url:
          type: string
          format: uri
        metadata:
          $ref: "#/components/schemas/ReportMetadata"
        output:
          type: string
          description: Base64-encoded output. Decoded content is limited to 4 MiB by default.
        created_at:
          type: string
          format: date-time
    ReportType:
      type: string
      enum:
        - sast
        - dast
        - sbom
        - code_coverage
        - custom
    ReportMetadata:
      type: object
      properties: {}
    ReportCreateRequest:
      type: object
      required:
        - id
        - type
        - status
      properties:
        id:
          type: string
          minLength: 1
        type:
          $ref: "#/components/schemas/ReportType"
        status:
          $ref: "#/components/schemas/JobStatus"
        location:
          type: string
        url:
          $ref: "#/components/schemas/URL"
        metadata:
          $ref: "#/components/schemas/ReportMetadata"
        output:
          type: string
          description: Base64-encoded output. Decoded content is limited to 4 MiB by default.
    ReportUpdateRequest:
      type: object
      required:
        - status
      properties:
        status:
          $ref: "#/components/schemas/JobStatus"
        metadata:
          $ref: "#/components/schemas/ReportMetadata"
        output:
          type: string
          description: Base64-encoded output. Decoded content is limited to 4 MiB by default.
    ReportDataResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Report"
    ReportListResponse:
      type: object
      required:
        - data
        - pagination
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Report"
        pagination:
          $ref: "#/components/schemas/Pagination"
...